flutter_app: fix work request cancelled warning
Previously, we tried to wait for tick() function to complete.  However,
the function immediately invokes another function via MethodChannel, and
it seems that flutter context would be confused by this kind of
recursive call.  The tick() function will never finish, and on next
tick(), the previous context will be forced cancelled.  This is why we
are seeing the strange error message.